requirements.txt
absl-py==0.2.2
astor==0.6.2
bleach==1.5.0
gast==0.2.0
grpcio==1.12.1
html5lib==0.9999999
Markdown==2.6.11
numpy==1.14.5
pillow==6.0.0
protobuf==3.6.1
six==1.11.0
tensorboard==1.12
tensorflow-gpu==1.12
termcolor==1.1.0
Werkzeug==0.14.1
